requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
supabase==1.2.0
psycopg2-binary==2.9.9
Flask-WTF==1.2.1
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
gunicorn==21.2.0
pytest==7.4.2